backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
langchain==0.2.16
langchain-community==0.2.16
sentence-transformers==3.0.1
pinecone-client==3.2.2
sqlalchemy==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9
httpx==0.27.0
python-dotenv==1.0.1
pypdf==4.3.1
=======
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.112.0
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.4.0
greenlet==3.5.0
groq==1.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.27.0
huggingface_hub==0.36.2
idna==3.13
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
MarkupSafe==3.0.3
marshmallow==3.26.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==1.26.4
orjson==3.11.8
packaging==24.2
pillow==12.2.0
pinecone-client==3.2.2
propcache==0.4.1
pydantic==2.13.3
pydantic_core==2.46.3
pypdf==4.3.1
python-dotenv==1.0.1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.0.1
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.31
starlette==0.37.2
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.7.0
tqdm==4.67.3
transformers==4.57.6
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.6
yarl==1.23.0
>>>>>>> 91f3e15 (save changes before pull)
